requirements.txt
# Kern
python-dotenv>=1.0.1
pydantic>=2.7
PyYAML>=6.0.1
requests>=2.32
httpx>=0.27

# Daten & Analyse
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
seaborn>=0.13

# CLI & Helfer
typer>=0.12
tqdm>=4.66
rich>=13.7

# API-Clients (Cloud-Modelle)
openai>=1.40
anthropic>=0.34
# xAI SDK für Grok (optional, aber empfohlen)
xai-sdk>=0.2
# Mistral API SDK
mistralai>=1.2.2

# Lokale Modelle (Teuken 7B)
torch>=2.3
transformers>=4.44
sentencepiece>=0.2
huggingface_hub>=0.24

# Google Gemini (Judge)
google-genai>=0.3.0
